requirements.txt
# 基本依赖
requests>=2.31.0
beautifulsoup4>=4.12.0
cloudscraper>=1.2.71
tqdm>=4.65.0

# PDF 处理
img2pdf>=0.4.4
Pillow>=10.0.0

# 数据库
mysql-connector-python>=8.0.0

# 命令行美化
pyfiglet>=1.0.0
rich>=13.7.0

# 其他工具
urllib3>=2.0.0
